support for custom input/output files
- making it no longer a requirement to download both a vtt and video using yt-dlp to use the vtt-parsing tools

+++ data_processing/parsed_to_tokens.py
@@ -64,10 +64,16 @@ def process_file(f) -> pd.DataFrame or None:
     return df
 
 
-def main():
+def main(custom_input_directory=None, custom_output_directory=None):
+    """
+    Processes the parsed data and generates a dataset containing timestamped tokens.
+    :param custom_input_directory: by default, will look at data/parsed/.
+    :param custom_output_directory: by default, will output to data/final_dataset/.
+    :return:
+    """
     data_path = "../data/"
-    input_path = data_path + "parsed/"
-    output_path = data_path + "final_dataset/csvs/"
+    input_path = custom_input_directory if custom_input_directory else data_path + "parsed/"
+    output_path = custom_output_directory if custom_output_directory else data_path + "final_dataset/"
     index_file_path = data_path + "indexed/index.tsv"
 
     for filename in os.listdir(os.path.abspath(input_path)):
@@ -78,17 +84,21 @@ def main():
                 continue
 
             # if the directory does not exist, create it
-            csv_output = output_path
+            csv_output = output_path + "/csvs/"
             if not os.path.exists(os.path.abspath(csv_output)):
                 os.makedirs(os.path.abspath(csv_output))
 
             df.to_csv(os.path.abspath(csv_output + filename), index=False)
 
-        # make a copy of the index file
-        with open(os.path.abspath(index_file_path)) as f:
-            df = pd.read_csv(f, sep='\t')
-            df.to_csv(os.path.abspath(output_path + "index.tsv"), sep='\t', index=False)
+        # index file
+        idx_path = os.path.abspath(index_file_path)
+        if os.path.exists(idx_path):
+            with open(idx_path) as f:
+                df = pd.read_csv(f, sep='\t')
+                df.to_csv(os.path.abspath(output_path + "index.tsv"), sep='\t', index=False)
+        else:
+            print("Index file does not exist. If this is intended, please ignore this message.")
 
 
 if __name__ == '__main__':
-    main()
+    main(custom_input_directory=None, custom_output_directory=None)
